+++ deepchem/models/__init__.py
@@ -51,6 +51,7 @@ try:
     from deepchem.models.torch_models import ResidueEmbedding
     from deepchem.models.torch_models import PositionalEncoding
     from deepchem.models.torch_models import CosineSchedule
+    from deepchem.models.torch_models import BackboneDiffusion
     from deepchem.models.torch_models import ScaledDotProductAttention, SelfAttention
     from deepchem.models.torch_models import GroverReadout
     from deepchem.models.torch_models import WeaveModel, DTNNModel
